refactor(enrollment): collapse Enrollment aggregate to schema-as-struct
Slice 3 of the Enrollment flatten — the aggregate root. Merge the domain
model + Ecto schema + mapper + repository + 2 ports into
KlassHero.Enrollment.Enrollment (schema + changesets + lifecycle transitions
+ predicates), and inline the create/confirm/cancel commands and all the
read queries onto the context, calling Repo directly. State-changing
operations are wrapped in context_span.

- create_with_capacity_check copied verbatim (SELECT ... FOR UPDATE race guard).
- Enriched roster/pending reads inline the Family/ProgramCatalog ACL joins;
  the raw programs-table query is preserved to avoid the ProgramCatalog cycle.
- Backpex admin booking LiveView, status filter, factory and seeds repointed
  to the collapsed module + admin_changeset/3.
- cancel_enrollment_by_admin guard relaxed so the ensure_reason_present path
  returns {:error, :invalid_reason} instead of crashing on empty input.
- Test tree flattened: command/query tests repointed to the public API and
  moved to test/klass_hero/enrollment/; enrollment_entity_test covers the
  changesets, transitions and predicates.
