+++ lib/klass_hero/enrollment.ex
@@ -10,16 +10,16 @@ defmodule KlassHero.Enrollment do
 
   import Ecto.Query, warn: false
 
+  alias KlassHero.Enrollment.Adapters.Driven.ACL.ChildInfoACL
+  alias KlassHero.Enrollment.Adapters.Driven.ACL.ParentInfoACL
   alias KlassHero.Enrollment.Adapters.Driven.ACL.ParticipantDetailsACL
+  alias KlassHero.Enrollment.Adapters.Driven.ACL.ProgramCatalogACL
   alias KlassHero.Enrollment.Adapters.Driven.ACL.ProgramScheduleACL
-  alias KlassHero.Enrollment.Adapters.Driven.Persistence.Schemas.EnrollmentSchema
+  alias KlassHero.Enrollment.Adapters.Driven.Persistence.Queries.EnrollmentQueries
   alias KlassHero.Enrollment.Adapters.Driving.Events.EventHandlers.NotifyLiveViews
 
   alias KlassHero.Enrollment.Application.Commands.{
-    CancelEnrollmentByAdmin,
     ClaimInvite,
-    ConfirmEnrollment,
-    CreateEnrollment,
     DeleteInvite,
     ImportEnrollmentCsv,
     InviteSingleParticipant,
@@ -29,24 +29,21 @@ defmodule KlassHero.Enrollment do
   alias KlassHero.Enrollment.Application.ParticipantPolicyForm
 
   alias KlassHero.Enrollment.Application.Queries.{
-    CheckEnrollment,
-    CountMonthlyBookings,
     CountProgramInvites,
-    GetBookingUsageInfo,
-    GetEnrollment,
-    ListEnrolledIdentityIds,
-    ListParentEnrollments,
-    ListPendingEnrollmentsForProvider,
-    ListProgramEnrollments,
     ListProgramInvites
   }
 
   alias KlassHero.Enrollment.Application.SingleInviteForm
   alias KlassHero.Enrollment.Domain.Events.EnrollmentEvents
   alias KlassHero.Enrollment.Domain.Services.EnrollmentClassifier
+  alias KlassHero.Enrollment.Enrollment
   alias KlassHero.Enrollment.EnrollmentPolicy
   alias KlassHero.Enrollment.ParticipantPolicy
+  alias KlassHero.Family
+  alias KlassHero.Family.ParentProfile
   alias KlassHero.Repo
+  alias KlassHero.Shared.Adapters.Driven.Persistence.EctoErrorHelpers
+  alias KlassHero.Shared.Entitlements
   alias KlassHero.Shared.EventDispatchHelper
 
   @active_statuses ~w(pending confirmed)
@@ -58,7 +55,83 @@ defmodule KlassHero.Enrollment do
   already exists for the child/program, or `{:error, term()}` on validation failure.
   """
   def create_enrollment(params) when is_map(params) do
-    CreateEnrollment.execute(params)
+    context_span entity: "enrollment" do
+      do_create_enrollment(params)
+    end
+  end
+
+  defp do_create_enrollment(%{identity_id: identity_id} = params) when is_binary(identity_id) do
+    with {:ok, parent} <- fetch_parent(identity_id),
+         {:ok, _parent} <-
+           Entitlements.ensure_booking_capacity(parent, count_monthly_bookings(parent.id)),
+         {:ok, :eligible} <- ensure_eligible(params[:program_id], params[:child_id]) do
+      params
+      |> build_enrollment_attrs(parent.id)
+      |> persist_and_dispatch(identity_id)
+    end
+  end
+
+  defp do_create_enrollment(params) do
+    params
+    |> build_enrollment_attrs(params[:parent_id])
+    |> persist_and_dispatch(params[:identity_id])
+  end
+
+  defp fetch_parent(identity_id) do
+    case Family.get_parent_by_identity(identity_id) do
+      {:ok, parent} -> {:ok, parent}
+      {:error, :not_found} -> {:error, :no_parent_profile}
+    end
+  end
+
+  # 3-tuple {:error, :ineligible, reasons} bubbles verbatim; 2-tuple lookup failures map to
+  # :processing_failed (fail-closed when eligibility cannot be verified).
+  defp ensure_eligible(program_id, child_id) do
+    case check_participant_eligibility(program_id, child_id) do
+      {:ok, :eligible} -> {:ok, :eligible}
+      {:error, :ineligible, reasons} -> {:error, :ineligible, reasons}
+      {:error, _reason} -> {:error, :processing_failed}
+    end
+  end
+
+  defp build_enrollment_attrs(params, parent_id) do
+    %{
+      program_id: params[:program_id],
+      child_id: params[:child_id],
+      parent_id: parent_id,
+      status: params[:status] || :pending,
+      enrolled_at: params[:enrolled_at] || DateTime.utc_now(),
+      subtotal: params[:subtotal],
+      vat_amount: params[:vat_amount],
+      card_fee_amount: params[:card_fee_amount],
+      total_amount: params[:total_amount],
+      payment_method: params[:payment_method],
+      special_requirements: params[:special_requirements]
+    }
+  end
+
+  defp persist_and_dispatch(attrs, identity_id) do
+    case create_enrollment_with_capacity_check(attrs, attrs[:program_id]) do
+      {:ok, enrollment} ->
+        dispatch_enrollment_created(enrollment, identity_id)
+        {:ok, enrollment}
+
+      error ->
+        error
+    end
+  end
+
+  # Fire-and-forget — a failed handler must not roll back a successful enrollment.
+  defp dispatch_enrollment_created(enrollment, identity_id) do
+    EnrollmentEvents.enrollment_created(enrollment.id, %{
+      enrollment_id: enrollment.id,
+      child_id: enrollment.child_id,
+      parent_id: enrollment.parent_id,
+      parent_user_id: identity_id,
+      program_id: enrollment.program_id,
+      status: enrollment.status
+    })
+    |> EventDispatchHelper.dispatch(__MODULE__)
   end
 
   @doc """
@@ -70,8 +143,30 @@ defmodule KlassHero.Enrollment do
   or `{:error, :invalid_reason}`.
   """
   def cancel_enrollment_by_admin(enrollment_id, admin_id, reason)
-      when is_binary(enrollment_id) and is_binary(admin_id) and is_binary(reason) and byte_size(reason) > 0 do
-    CancelEnrollmentByAdmin.execute(enrollment_id, admin_id, reason)
+      when is_binary(enrollment_id) and is_binary(admin_id) do
+    context_span entity: "enrollment" do
+      with {:ok, reason} <- Enrollment.ensure_reason_present(reason),
+           {:ok, enrollment} <- get_enrollment(enrollment_id),
+           {:ok, cancelled} <- Enrollment.cancel(enrollment, reason),
+           {:ok, persisted} <-
+             update_enrollment(enrollment_id, %{
+               status: cancelled.status,
+               cancelled_at: cancelled.cancelled_at,
+               cancellation_reason: cancelled.cancellation_reason
+             }) do
+        persisted.id
+        |> EnrollmentEvents.enrollment_cancelled(%{
+          enrollment_id: persisted.id,
+          program_id: persisted.program_id,
+          child_id: persisted.child_id,
+          parent_id: persisted.parent_id,
+          admin_id: admin_id,
+          reason: reason,
+          cancelled_at: persisted.cancelled_at
+        })
+        |> EventDispatchHelper.dispatch_or_ok(__MODULE__, persisted)
+      end
+    end
   end
 
   @doc """
@@ -80,9 +175,55 @@ defmodule KlassHero.Enrollment do
   Returns `{:ok, Enrollment.t()}`, `{:error, :not_found}`, `{:error, :unauthorized}`,
   or `{:error, :invalid_status_transition}`.
   """
-  def confirm_enrollment(%{enrollment_id: enrollment_id, provider_id: provider_id} = params)
+  def confirm_enrollment(%{enrollment_id: enrollment_id, provider_id: provider_id})
       when is_binary(enrollment_id) and is_binary(provider_id) do
-    ConfirmEnrollment.execute(params)
+    context_span entity: "enrollment" do
+      with {:ok, enrollment_id} <- cast_uuid_or_not_found(enrollment_id),
+           {:ok, enrollment} <- get_enrollment(enrollment_id),
+           :ok <- authorize_provider(enrollment, provider_id),
+           {:ok, confirmed} <- Enrollment.confirm(enrollment),
+           {:ok, persisted} <-
+             update_enrollment(enrollment_id, %{status: confirmed.status, confirmed_at: confirmed.confirmed_at}) do
+        dispatch_confirmation_event(persisted, provider_id)
+      end
+    end
+  end
+
+  # `enrollment_id` arrives from the LiveView `phx-value-id` (DOM-tamperable). Without this
+  # guard, `Repo.get/2` raises `Ecto.Query.CastError` on malformed input. `provider_id` is
+  # server-trusted (read from `current_scope`).
+  defp cast_uuid_or_not_found(id) do
+    case Ecto.UUID.cast(id) do
+      {:ok, _} -> {:ok, id}
+      :error -> {:error, :not_found}
+    end
+  end
+
+  defp authorize_provider(%Enrollment{program_id: program_id}, provider_id) do
+    if ProgramCatalogACL.program_owned_by?(program_id, provider_id) do
+      :ok
+    else
+      {:error, :unauthorized}
+    end
+  end
+
+  defp dispatch_confirmation_event(%Enrollment{} = persisted, provider_id) do
+    dispatch_result =
+      persisted.id
+      |> EnrollmentEvents.enrollment_confirmed(%{
+        enrollment_id: persisted.id,
+        program_id: persisted.program_id,
+        provider_id: provider_id,
+        child_id: persisted.child_id,
+        parent_id: persisted.parent_id,
+        confirmed_at: persisted.confirmed_at
+      })
+      |> EventDispatchHelper.dispatch_or_error(__MODULE__)
+
+    case dispatch_result do
+      :ok -> {:ok, persisted}
+      {:error, _} = err -> err
+    end
   end
 
   @doc """
@@ -232,14 +373,20 @@ defmodule KlassHero.Enrollment do
   Retrieves an enrollment by ID. Returns `{:ok, Enrollment.t()}` or `{:error, :not_found}`.
   """
   def get_enrollment(id) when is_binary(id) do
-    GetEnrollment.execute(id)
+    case Repo.get(Enrollment, id) do
+      nil -> {:error, :not_found}
+      enrollment -> {:ok, enrollment}
+    end
   end
 
   @doc """
   Lists all enrollments for a parent, ordered by `enrolled_at` descending.
   """
   def list_parent_enrollments(parent_id) when is_binary(parent_id) do
-    ListParentEnrollments.execute(parent_id)
+    EnrollmentQueries.base()
+    |> EnrollmentQueries.by_parent(parent_id)
+    |> EnrollmentQueries.order_by_enrolled_at_desc()
+    |> Repo.all()
   end
 
   @doc """
@@ -261,7 +408,64 @@ defmodule KlassHero.Enrollment do
   Used by the provider dashboard to display the program roster.
   """
   def list_program_enrollments(program_id) when is_binary(program_id) do
-    ListProgramEnrollments.execute(program_id)
+    case list_active_by_program(program_id) do
+      [] ->
+        []
+
+      enrollments ->
+        child_map = child_map_for(enrollments)
+        parent_map = parent_map_for(enrollments)
+        Enum.map(enrollments, &build_roster_entry(&1, child_map, parent_map))
+    end
+  end
+
+  defp list_active_by_program(program_id) do
+    EnrollmentQueries.base()
+    |> EnrollmentQueries.by_program(program_id)
+    |> EnrollmentQueries.active_only()
+    |> EnrollmentQueries.order_by_enrolled_at_desc()
+    |> Repo.all()
+  end
+
+  defp child_map_for(enrollments) do
+    enrollments
+    |> Enum.map(& &1.child_id)
+    |> Enum.uniq()
+    |> ChildInfoACL.get_children_by_ids()
+    |> Map.new(fn c -> {c.id, c} end)
+  end
+
+  defp parent_map_for(enrollments) do
+    enrollments
+    |> Enum.map(& &1.parent_id)
+    |> Enum.uniq()
+    |> ParentInfoACL.get_parents_by_ids()
+    |> Map.new(fn p -> {p.id, p} end)
+  end
+
+  defp build_roster_entry(enrollment, child_map, parent_map) do
+    child_name =
+      case Map.get(child_map, enrollment.child_id) do
+        nil -> "Unknown"
+        child -> "#{child.first_name} #{child.last_name}"
+      end
+
+    # nil parent_user_id disables the message button in UI (orphaned/deleted parent profile)
+    parent_user_id =
+      case Map.get(parent_map, enrollment.parent_id) do
+        nil -> nil
+        parent -> parent.identity_id
+      end
+
+    %{
+      enrollment_id: enrollment.id,
+      child_id: enrollment.child_id,
+      child_name: child_name,
+      parent_id: enrollment.parent_id,
+      parent_user_id: parent_user_id,
+      status: enrollment.status,
+      enrolled_at: enrollment.enrolled_at
+    }
   end
 
   @doc """
@@ -270,8 +474,61 @@ defmodule KlassHero.Enrollment do
   Used by the provider dashboard's "Pending enrollments" inbox card to
   surface enrollments awaiting provider approval.
   """
+  def list_pending_enrollments_for_provider([]), do: []
+
   def list_pending_enrollments_for_provider(program_ids) when is_list(program_ids) do
-    ListPendingEnrollmentsForProvider.execute(program_ids)
+    case list_pending_by_programs(program_ids) do
+      [] ->
+        []
+
+      enrollments ->
+        child_map = child_map_for(enrollments)
+        program_map = program_titles_map(program_ids)
+        Enum.map(enrollments, &build_pending_entry(&1, child_map, program_map))
+    end
+  end
+
+  defp list_pending_by_programs(program_ids) do
+    Enrollment
+    |> where([e], e.status == :pending and e.program_id in ^program_ids)
+    |> Repo.all()
+  end
+
+  defp program_titles_map(program_ids) do
+    # Querying `programs` directly avoids a ProgramCatalog↔Enrollment dependency cycle
+    # (ProgramCatalog already depends on Enrollment for capacity ACL).
+    valid_ids = Enum.filter(program_ids, fn id -> match?({:ok, _}, Ecto.UUID.cast(id)) end)
+
+    case valid_ids do
+      [] ->
+        %{}
+
+      ids ->
+        from(p in "programs",
+          where: p.id in ^Enum.map(ids, &Ecto.UUID.dump!/1),
+          select: {type(p.id, :binary_id), p.title}
+        )
+        |> Repo.all()
+        |> Map.new()
+    end
+  end
+
+  defp build_pending_entry(enrollment, child_map, program_map) do
+    child_name =
+      case Map.get(child_map, enrollment.child_id) do
+        nil -> "Unknown"
+        child -> "#{child.first_name} #{child.last_name}"
+      end
+
+    %{
+      enrollment_id: enrollment.id,
+      program_id: enrollment.program_id,
+      program_title: Map.get(program_map, enrollment.program_id, "Unknown"),
+      child_id: enrollment.child_id,
+      child_name: child_name,
+      parent_id: enrollment.parent_id,
+      enrolled_at: enrollment.enrolled_at
+    }
   end
 
   @doc """
@@ -287,7 +544,16 @@ defmodule KlassHero.Enrollment do
   Used by the entitlements system to enforce monthly booking limits.
   """
   def count_monthly_bookings(parent_id, month \\ nil) when is_binary(parent_id) do
-    CountMonthlyBookings.execute(parent_id, month)
+    date = month || Date.utc_today()
+    start_date = Date.beginning_of_month(date)
+    end_date = Date.end_of_month(date)
+
+    EnrollmentQueries.base()
+    |> EnrollmentQueries.by_parent(parent_id)
+    |> EnrollmentQueries.active_only()
+    |> EnrollmentQueries.by_date_range(start_date, end_date)
+    |> EnrollmentQueries.count()
+    |> Repo.one()
   end
 
   @doc """
@@ -295,7 +561,23 @@ defmodule KlassHero.Enrollment do
   (`parent_id`, `tier`, `cap`, `used`, `remaining`), or `{:error, :no_parent_profile}`.
   """
   def get_booking_usage_info(identity_id) when is_binary(identity_id) do
-    GetBookingUsageInfo.execute(identity_id)
+    case Family.get_parent_by_identity(identity_id) do
+      {:ok, parent} -> {:ok, build_usage_info(parent)}
+      {:error, :not_found} -> {:error, :no_parent_profile}
+    end
+  end
+
+  defp build_usage_info(parent) do
+    cap = Entitlements.monthly_booking_cap(parent)
+    used = count_monthly_bookings(parent.id)
+
+    remaining =
+      case cap do
+        :unlimited -> :unlimited
+        cap -> max(0, cap - used)
+      end
+
+    %{parent_id: parent.id, tier: parent.subscription_tier, cap: cap, used: used, remaining: remaining}
   end
 
   @doc """
@@ -304,7 +586,13 @@ defmodule KlassHero.Enrollment do
   """
   @spec list_enrolled_identity_ids(String.t()) :: [String.t()]
   def list_enrolled_identity_ids(program_id) when is_binary(program_id) do
-    ListEnrolledIdentityIds.execute(program_id)
+    EnrollmentQueries.base()
+    |> EnrollmentQueries.by_program(program_id)
+    |> EnrollmentQueries.active_only()
+    |> join(:inner, [e], p in ParentProfile, on: e.parent_id == p.id)
+    |> select([e, p], p.identity_id)
+    |> distinct(true)
+    |> Repo.all()
   end
 
   @doc """
@@ -314,7 +602,87 @@ defmodule KlassHero.Enrollment do
   """
   @spec enrolled?(String.t(), String.t()) :: boolean()
   def enrolled?(program_id, identity_id) when is_binary(program_id) and is_binary(identity_id) do
-    CheckEnrollment.execute(program_id, identity_id)
+    EnrollmentQueries.base()
+    |> EnrollmentQueries.by_program(program_id)
+    |> EnrollmentQueries.active_only()
+    |> join(:inner, [e], p in ParentProfile, on: e.parent_id == p.id)
+    |> where([e, p], p.identity_id == ^identity_id)
+    |> Repo.exists?()
+  end
+
+  # Creates an enrollment with an atomic capacity check. Locks the enrollment policy row
+  # (`SELECT FOR UPDATE`) inside a transaction to prevent TOCTOU races where concurrent
+  # requests could both pass the capacity check. Falls through to a plain insert when
+  # program_id is nil (no policy to lock).
+  defp create_enrollment_with_capacity_check(attrs, nil), do: create_enrollment_record(attrs)
+
+  defp create_enrollment_with_capacity_check(attrs, program_id) when is_map(attrs) and is_binary(program_id) do
+    Ecto.Multi.new()
+    |> Ecto.Multi.run(:lock_and_check, fn repo, _changes ->
+      query = from(p in EnrollmentPolicy, where: p.program_id == ^program_id, lock: "FOR UPDATE")
+
+      case repo.one(query) do
+        nil ->
+          {:ok, :unlimited}
+
+        %EnrollmentPolicy{} = policy ->
+          active = count_active_enrollments_in_tx(repo, program_id)
+          check_capacity(policy, active)
+      end
+    end)
+    |> Ecto.Multi.run(:create, fn _repo, _changes -> create_enrollment_record(attrs) end)
+    |> Repo.transaction()
+    |> case do
+      {:ok, %{create: enrollment}} -> {:ok, enrollment}
+      {:error, :lock_and_check, :program_full, _} -> {:error, :program_full}
+      {:error, :create, reason, _} -> {:error, reason}
+    end
+  end
+
+  defp check_capacity(policy, active) do
+    if EnrollmentPolicy.has_capacity?(policy, active) do
+      remaining = if policy.max_enrollment, do: policy.max_enrollment - active, else: :unlimited
+      {:ok, remaining}
+    else
+      {:error, :program_full}
+    end
+  end
+
+  defp count_active_enrollments_in_tx(repo, program_id) do
+    from(e in Enrollment,
+      where: e.program_id == ^program_id and e.status in ^@active_statuses,
+      select: count(e.id)
+    )
+    |> repo.one()
+  end
+
+  defp create_enrollment_record(attrs) do
+    %Enrollment{}
+    |> Enrollment.create_changeset(attrs)
+    |> Repo.insert()
+    |> case do
+      {:ok, enrollment} ->
+        {:ok, enrollment}
+
+      {:error, %Ecto.Changeset{errors: errors} = changeset} ->
+        if EctoErrorHelpers.unique_constraint_violation?(errors, :program_id) do
+          {:error, :duplicate_resource}
+        else
+          {:error, changeset}
+        end
+    end
+  end
+
+  defp update_enrollment(id, attrs) do
+    case Repo.get(Enrollment, id) do
+      nil ->
+        {:error, :not_found}
+
+      enrollment ->
+        enrollment
+        |> Enrollment.update_changeset(attrs)
+        |> Repo.update()
+    end
   end
 
   @doc """
@@ -365,7 +733,7 @@ defmodule KlassHero.Enrollment do
   Returns the count of active (pending/confirmed) enrollments for a program.
   """
   def count_active_enrollments(program_id) when is_binary(program_id) do
-    from(e in EnrollmentSchema,
+    from(e in Enrollment,
       where: e.program_id == ^program_id and e.status in ^@active_statuses,
       select: count(e.id)
     )
@@ -380,7 +748,7 @@ defmodule KlassHero.Enrollment do
 
   def count_active_enrollments_batch(program_ids) when is_list(program_ids) do
     counts =
-      from(e in EnrollmentSchema,
+      from(e in Enrollment,
         where: e.program_id in ^program_ids and e.status in ^@active_statuses,
         group_by: e.program_id,
         select: {e.program_id, count(e.id)}
