--- lib/klass_hero/enrollment/application/queries/get_enrollment.ex
@@ -1,27 +1,0 @@
-defmodule KlassHero.Enrollment.Application.Queries.GetEnrollment do
-  @moduledoc """
-  Use case for retrieving an enrollment by ID.
-  """
-
-  alias KlassHero.Enrollment.Domain.Models.Enrollment
-
-  require Logger
-
-  @enrollment_repository Application.compile_env!(:klass_hero, [
-                           :enrollment,
-                           :for_querying_enrollments
-                         ])
-
-  @doc """
-  Retrieves an enrollment by its ID.
-
-  Returns:
-  - `{:ok, Enrollment.t()}` when found
-  - `{:error, :not_found}` when no enrollment exists with the given ID
-  """
-  @spec execute(binary()) :: {:ok, Enrollment.t()} | {:error, :not_found}
-  def execute(id) when is_binary(id) do
-    Logger.debug("[Enrollment.GetEnrollment] Fetching enrollment", enrollment_id: id)
-    @enrollment_repository.get_by_id(id)
-  end
-end
