+++ test/klass_hero/enrollment/cancel_enrollment_by_admin_test.exs
@@ -1,10 +1,9 @@
-defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest do
+defmodule KlassHero.Enrollment.CancelEnrollmentByAdminTest do
   use KlassHero.DataCase, async: true
 
   import KlassHero.Factory
 
-  alias KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdmin
-  alias KlassHero.Enrollment.Domain.Models.Enrollment
+  alias KlassHero.Enrollment.Enrollment
 
   describe "execute/3" do
     test "cancels a pending enrollment and returns domain entity" do
@@ -12,7 +11,7 @@ defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest 
       admin_id = Ecto.UUID.generate()
 
       assert {:ok, enrollment} =
-               CancelEnrollmentByAdmin.execute(schema.id, admin_id, "Duplicate booking")
+               KlassHero.Enrollment.cancel_enrollment_by_admin(schema.id, admin_id, "Duplicate booking")
 
       assert %Enrollment{} = enrollment
       assert enrollment.status == :cancelled
@@ -25,7 +24,7 @@ defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest 
       admin_id = Ecto.UUID.generate()
 
       assert {:ok, enrollment} =
-               CancelEnrollmentByAdmin.execute(schema.id, admin_id, "Parent requested")
+               KlassHero.Enrollment.cancel_enrollment_by_admin(schema.id, admin_id, "Parent requested")
 
       assert enrollment.status == :cancelled
     end
@@ -35,7 +34,7 @@ defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest 
       admin_id = Ecto.UUID.generate()
 
       assert {:error, :invalid_status_transition} =
-               CancelEnrollmentByAdmin.execute(schema.id, admin_id, "Too late")
+               KlassHero.Enrollment.cancel_enrollment_by_admin(schema.id, admin_id, "Too late")
     end
 
     test "returns invalid_status_transition for already cancelled enrollment" do
@@ -43,14 +42,14 @@ defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest 
       admin_id = Ecto.UUID.generate()
 
       assert {:error, :invalid_status_transition} =
-               CancelEnrollmentByAdmin.execute(schema.id, admin_id, "Already gone")
+               KlassHero.Enrollment.cancel_enrollment_by_admin(schema.id, admin_id, "Already gone")
     end
 
     test "returns not_found for nonexistent enrollment" do
       admin_id = Ecto.UUID.generate()
 
       assert {:error, :not_found} =
-               CancelEnrollmentByAdmin.execute(Ecto.UUID.generate(), admin_id, "Nope")
+               KlassHero.Enrollment.cancel_enrollment_by_admin(Ecto.UUID.generate(), admin_id, "Nope")
     end
 
     test "returns invalid_reason for empty reason" do
@@ -58,7 +57,7 @@ defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest 
       admin_id = Ecto.UUID.generate()
 
       assert {:error, :invalid_reason} =
-               CancelEnrollmentByAdmin.execute(schema.id, admin_id, "")
+               KlassHero.Enrollment.cancel_enrollment_by_admin(schema.id, admin_id, "")
     end
 
     test "returns invalid_reason for nil reason" do
@@ -66,7 +65,7 @@ defmodule KlassHero.Enrollment.Application.Commands.CancelEnrollmentByAdminTest 
       admin_id = Ecto.UUID.generate()
 
       assert {:error, :invalid_reason} =
-               CancelEnrollmentByAdmin.execute(schema.id, admin_id, nil)
+               KlassHero.Enrollment.cancel_enrollment_by_admin(schema.id, admin_id, nil)
     end
   end
 end
